requirements.txt
gradio
Pillow
transformers
datasets
torch
soundfile
numpy
tensorflow-cpu
huggingface_hub
numpy
SentencePiece